requirements/local.txt
# dev and test
# ------------
pytest==7.4.3
pytest_mock==3.12.0

# Code linters, formatters, and security scanners
# ------------
black==23.11.0
flake8==6.1.0
flake8-coding==1.3.2
pre-commit==3.5.0
isort==5.12.0
mypy==1.7.1
pylint==3.0.2
bandit==1.7.5
pydocstringformatter==0.7.3
tox==4.11.4
codespell==2.2.6

# project dependencies
# ------------
python-decouple==3.8
